# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Playona API
  description: 크로스 플랫폼 음악 공유 서비스 API
  version: 1.0.2

servers:
  - url: http://localhost:8080
    description: 로컬 개발 서버
  - url: https://api.playona.io
    description: 운영 서버

tags:
  - name: 인증
    description: OAuth 로그인 및 토큰 관리
  - name: 유저
    description: 유저 정보 및 선호 플랫폼
  - name: 플랫폼
    description: 지원 플랫폼 목록
  - name: 트랙
    description: 곡 정보 조회 및 플랫폼 매칭
  - name: 공유링크
    description: 공유 링크 생성 및 리디렉션

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        message:
          type: string
          nullable: true
          example: null
        result:
          type: string
          example: SUCCESS

    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
          example: eyJhbGciOiJIUzI1NiJ9...
        refreshToken:
          type: string
          example: eyJhbGciOiJIUzI1NiJ9...

    UserResponse:
      type: object
      properties:
        userUuid:
          type: string
          example: 550e8400-e29b-41d4-a716
        email:
          type: string
          example: user@example.com
        nickname:
          type: string
          example: 윤성
        profileImageUrl:
          type: string
          example: https://k.kakaocdn.net/...
        createdAt:
          type: string
          format: date-time

    PlatformResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Spotify
        slug:
          type: string
          example: spotify
        logoUrl:
          type: string
          nullable: true
          example: null
        isActive:
          type: boolean
          example: true

    PlatformPreferenceResponse:
      type: object
      properties:
        platformId:
          type: integer
          example: 1
        platformName:
          type: string
          example: Spotify
        priority:
          type: integer
          example: 1

    PlatformUrlResponse:
      type: object
      properties:
        slug:
          type: string
          example: spotify
        name:
          type: string
          example: Spotify
        url:
          type: string
          example: https://open.spotify.com/track/abc123
        thumbnailUrl:
          type: string
          example: https://i.scdn.co/image/...

    SharedLinkResponse:
      type: object
      properties:
        shortCode:
          type: string
          example: abc123
        trackTitle:
          type: string
          example: Blinding Lights
        trackArtist:
          type: string
          example: The Weeknd
        thumbnailUrl:
          type: string
          example: https://i.scdn.co/image/...
        clickCount:
          type: integer
          example: 42
        shareUrl:
          type: string
          example: http://localhost:8080/t/abc123
        platforms:
          type: array
          items:
            $ref: '#/components/schemas/PlatformUrlResponse'

    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        data:
          type: object
          nullable: true
          example: null
        message:
          type: string
          example: 곡을 찾을 수 없습니다
        result:
          type: string
          example: FAIL

paths:
  # ========== 인증 ==========
  /oauth2/authorization/kakao:
    get:
      tags: [인증]
      summary: 카카오 로그인
      description: |
        카카오 OAuth2 로그인 페이지로 리다이렉트합니다.
        로그인 성공 시 `{OAUTH2_REDIRECT_URL}?token={accessToken}&refresh={refreshToken}` 으로 리다이렉트됩니다.
      responses:
        '302':
          description: 카카오 로그인 페이지로 리다이렉트

  /oauth2/authorization/google:
    get:
      tags: [인증]
      summary: 구글 로그인
      description: |
        구글 OAuth2 로그인 페이지로 리다이렉트합니다.
        로그인 성공 시 `{OAUTH2_REDIRECT_URL}?token={accessToken}&refresh={refreshToken}` 으로 리다이렉트됩니다.
      responses:
        '302':
          description: 구글 로그인 페이지로 리다이렉트

  /api/auth/refresh:
    post:
      tags: [인증]
      summary: 토큰 재발급
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
                  example: eyJhbGci...
      responses:
        '200':
          description: 재발급 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Refresh Token 만료 또는 유효하지 않음

  /api/auth/logout:
    post:
      tags: [인증]
      summary: 로그아웃
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: 로그아웃 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # ========== 유저 ==========
  /api/users/me:
    get:
      tags: [유저]
      summary: 내 정보 조회
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserResponse'
        '401':
          description: 인증 필요

    put:
      tags: [유저]
      summary: 내 정보 수정
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nickname:
                  type: string
                  example: 윤성
                profileImageUrl:
                  type: string
      responses:
        '200':
          description: 수정 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserResponse'

  /api/users/me/platforms:
    get:
      tags: [유저]
      summary: 선호 플랫폼 목록 조회
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/PlatformPreferenceResponse'

    put:
      tags: [유저]
      summary: 선호 플랫폼 설정
      description: 플랫폼 우선순위를 설정합니다 (1순위부터 순서대로)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                properties:
                  platformId:
                    type: integer
                    example: 1
                  priority:
                    type: integer
                    example: 1
      responses:
        '200':
          description: 설정 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/PlatformPreferenceResponse'

  # ========== 플랫폼 ==========
  /api/platforms:
    get:
      tags: [플랫폼]
      summary: 지원 플랫폼 목록 조회
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/PlatformResponse'

  # ========== 트랙 ==========
  /api/tracks/resolve:
    get:
      tags: [트랙]
      summary: URL로 곡 정보 조회
      parameters:
        - name: url
          in: query
          required: true
          schema:
            type: string
            example: https://open.spotify.com/track/abc123
      responses:
        '200':
          description: 조회 성공
        '400':
          description: 지원하지 않는 URL 형식

  /api/tracks/{trackId}:
    get:
      tags: [트랙]
      summary: 곡 상세 조회
      parameters:
        - name: trackId
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: 조회 성공
        '404':
          description: 곡 없음

  # ========== 공유 링크 ==========
  /api/links:
    post:
      tags: [공유링크]
      summary: 공유 링크 생성
      description: |
        플랫폼 URL을 입력하면 Playona 단축 링크를 생성합니다.
        - 지원 플랫폼: Spotify, YouTube Music, YouTube
        - 비로그인도 가능
        - 링크 생성 시 모든 활성 플랫폼에 대해 자동 매칭
      security:
        - bearerAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
                  example: https://open.spotify.com/track/abc123
      responses:
        '200':
          description: 링크 생성 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SharedLinkResponse'
        '400':
          description: 잘못된 URL 형식 또는 지원하지 않는 플랫폼
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/links/my:
    get:
      tags: [공유링크]
      summary: 내 링크 목록 조회
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/SharedLinkResponse'
        '401':
          description: 인증 필요

  /api/links/{shortCode}:
    get:
      tags: [공유링크]
      summary: 공유 링크 조회
      parameters:
        - name: shortCode
          in: path
          required: true
          schema:
            type: string
            example: abc123
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SharedLinkResponse'
        '404':
          description: 링크 없음

  /api/links/{shortCode}/redirect:
    get:
      tags: [공유링크]
      summary: 선호 플랫폼으로 리디렉션
      description: |
        로그인 유저의 선호 플랫폼 순서대로 매칭된 URL로 리디렉션합니다.
        - 로그인 유저: 설정된 선호 플랫폼 1순위부터 매칭된 플랫폼으로 자동 리디렉션
        - 비로그인 유저: GET /api/links/{shortCode}/platforms 를 통해 플랫폼 선택 후 직접 이동
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: shortCode
          in: path
          required: true
          schema:
            type: string
            example: abc123
      responses:
        '302':
          description: 플랫폼 URL로 리디렉션
        '404':
          description: 링크 없음

  /api/links/{shortCode}/platforms:
    get:
      tags: [공유링크]
      summary: 링크의 플랫폼별 URL 목록 조회
      description: 비로그인 유저가 플랫폼을 선택할 수 있도록 매칭된 플랫폼 URL 목록을 반환합니다.
      parameters:
        - name: shortCode
          in: path
          required: true
          schema:
            type: string
            example: abc123
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/PlatformUrlResponse'
        '404':
          description: 링크 없음